--- src/component/Atoms/Title/index.stories.jsx
@@ -1,4 +1,0 @@
-import { Title } from "C:/Users/sugi2/OneDrive/プログラミング/Ryu-frontend-basic-2024-todoapp/src/component/Atoms/Title/index.jsx";
-export default { component: Title };
-
-export const Default = {};
