+++ src/components/Atoms/Title/index.stories.jsx
@@ -1,0 +1,5 @@
+import { Title } from "./index";
+
+export default { component: Title };
+
+export const Default = {};
